+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog for package bob_audio
 
 0.1.1 (2026-03-07)
 ------------------
-* Add bidirectional conversion to convert node (ROS to FIFO support)
+* Add bidirectional conversion (FIFO/stdin <-> ROS) with exclusive mode selection
 * Contributors: Bob Ros
 
 0.1.0 (2026-03-05)